requirements.txt
appnope==0.1.0
decorator==4.0.4
Django==1.8.4
django-crispy-forms==1.5.2
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
Pillow==3.0.0
psycopg2==2.6.1
selenium==2.47.3
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
